Refactored the Node and Tree classes
Refactored the Node and Tree classes to use the Double type.
This will allow to use the compare methods at the node class and avoids
the insert, delete and search tree methods to fail.
Used the Double class since this is the most generic Number type.
Added the constructors for the node accepting all the primitive numeric
types (int, double, float, short, long, byte)
Added some Node tests and Tree tests as well